fix: Adds .impl property to Postgres INTERVAL
The missing `.impl` attribute causes any kind of filtering on an `interval`
column to fail.

`.impl` is called in `sql.sqltypes._AbstractInterval.coerce_compared_value`,
and when the `ColumnElement` is `sql.sqltypes.Interval` this will point to
`sql.sqltypes.DateTime`.

A `.impl` property on `INTERVAL` is used to create a path to `DateTime` so that
`INTERVAL` behaves more like `Interval`.

Fixes: #6649